+++ src/components/input-field/input-field.types.ts
@@ -11,4 +11,5 @@ export type InputType =
     | 'textarea'
     | 'time'
     | 'url'
+    | 'urlAsText'
     | 'week';