+++ src/components/menu-surface/menu-surface.tsx
@@ -41,7 +41,6 @@ export class MenuSurface {
     private host: HTMLLimelMenuSurfaceElement;
 
     private menuSurface: MDCMenuSurface;
-    private observer: IResizeObserver;
 
     constructor() {
         this.handleDocumentClick = this.handleDocumentClick.bind(this);
@@ -95,13 +94,6 @@ export class MenuSurface {
         window.addEventListener('resize', this.handleResize, {
             passive: true,
         });
-
-        if ('ResizeObserver' in window) {
-            const observer = new ResizeObserver(() => {
-                this.ensureMenuFitsInViewPort();
-            });
-            observer.observe(this.host);
-        }
     }
 
     private teardown() {
@@ -111,10 +103,6 @@ export class MenuSurface {
         });
         this.host.removeEventListener('keydown', this.handleKeyDown);
         window.removeEventListener('resize', this.handleResize);
-
-        if (this.observer) {
-            this.observer.unobserve(this.host);
-        }
     }
 
     private handleDocumentClick(event) {
@@ -175,31 +163,5 @@ export class MenuSurface {
             event.stopPropagation();
             this.dismiss.emit();
         }
-    }
-
-    private ensureMenuFitsInViewPort() {
-        this.host.style.height = 'auto';
-        setTimeout(() => {
-            const viewHeight = Math.max(
-                document.documentElement.clientHeight || 0,
-                window.innerHeight || 0
-            );
-            const { top, bottom } = this.host.getBoundingClientRect();
-            if (viewHeight > bottom && top > 0) {
-                // The surface is rendered inside the viewport :)
-                return;
-            }
-
-            // Set the height of the surface so that it fits either above or
-            // below the trigger, depending on where there is more space.
-            const spaceAboveTopOfSurface = Math.max(top, 0);
-            const spaceBelowTopOfSurface = Math.max(viewHeight - top, 0);
-            const extraCosmeticSpace = 16;
-            const maxHeight =
-                viewHeight -
-                Math.min(spaceAboveTopOfSurface, spaceBelowTopOfSurface) -
-                extraCosmeticSpace;
-            this.host.style.height = `${maxHeight}px`;
-        });
     }
 }