+++ src/components/popover/examples/popover.tsx
@@ -22,30 +22,7 @@ export class PopoverExample {
                 onClick={this.openPopover}
             />,
             <limel-popover open={this.isOpen} onClose={this.onPopoverClose}>
-                <span slot="header">
-                    <h3
-                        style={{
-                            margin: '0.5rem 1rem',
-                            color: 'rgb(var(--contrast-100))',
-                        }}
-                    >
-                        Header content
-                    </h3>
-                </span>
-                <span slot="body">
-                    <p style={{ margin: '0.5rem 1rem' }}>Body content</p>
-                </span>
-                <span slot="footer">
-                    <span
-                        style={{
-                            margin: '0.5rem 1rem',
-                            'font-size': '0.75rem',
-                            color: 'rgb(var(--contrast-800))',
-                        }}
-                    >
-                        Footer content
-                    </span>
-                </span>
+                <p style={{ margin: '0.5rem 1rem' }}>Content</p>
             </limel-popover>,
         ];
     }